Add pause to execution for better user experience
When running the program it starts running the default execution plan
without warning. This can cause a bad user experience if they are
unsure of what to expect.

Adding a small wait step to ensure there is opportunity to cancel
and review what will be executed before changes are made

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>