Fix a bug in the edge-auth template
Prior to this fix, the cookie for OFC auth was being set at the
wrong level i.e. the root domain, but needed to be set one level
down at the system level, so that user functions could not read
the value.

This has been patched on the community cluster already.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>