Add merge support for YAML files in apply command
This patch allows multiple override files to be specified that
will be merged into each parameter given, so you can use a base
file and then a series of overrides if desired.

Tested with unit tests. Inspiration for the feature comes from
@csakshaug who proposed the idea and found the merge library.

The merge library is used in Docker, K8s and other high-profile
projects.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>