Capture GitHub app via new command
This command allows users to create and capture the webhook
secret, private key (PEM), and GitHub App ID from the GitHub's
App creation URL.

A YAML overide file can be generated for users to merge with
their own or the base example.init.yaml.

Tested with my own account and Safari - the results were as
expected - correct URLs, secrets, and names in the GitHub site.

We may want to add omitempty attributes to the schema to prevent
elements and nodes being emitted with nil/empty values.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>