Change the order of checking cert-manager is ready before moving on
We currently install cert-manager, then just install openfaas chart
then we check if cert-manager is ready. This leads to openfaas failing
to install because the webhooks for cert-manager don't exist and helm
fails if any of the endpoints are not ready. See issue:
https://github.com/helm/helm/issues/6361

This has been tested by running before changing the code, and openfaas
failed to install. Then I changed the order and ran against a new cluster.
This install worked as expected.

Signed-off-by: Alistair Hey <alistair@heyal.co.uk>