Switch to using helm3 for bootstrap
There's an ongoing issue which has been patched in helm 3 but
which is not yet patched in helm 2. We also have a backlog item
to move to the secure by default helm 3.

This change also downloads helm 3 and uses a newer version of
go-execute which has a fix for setting the $PATH correctly to
access binaries we have downloaded.

* Use faas-cli to get kubeseal
* Apply secrets in re-runnable way in bash scripts to fix #184
* Update for local usage of helm3 path
* Also patches minio to use helm3-style argument for install
* PATH is now set for the process before invoking execTask so that
the helm binary can be found.
* Use rollout status for sealedsecrets - this is quicker and uses
less forking than a loop with an outer check.
* Update deps
* Use non-root user for Dockerfile
* Add note for k3d without Traefik
* Use helm3 to install cert-manager
* Add k3sup to Gopkg.toml
* Remove reset scripts - the reset scripts became a source of
confusion for the community and due to the amount of software
installed, it's hard to truely reset a cluster and purge all
API extensions / charts.

Fixes: #178
Fixes: #181
Fixes: #180
Closes: #183
Fixes: #184

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>