Merge the list of secrets
This patch adds a custom piece of unit-tested code to merge the
list of secrets using a union since the mergo library doesn't
appear to be able to do that at present.

Ref: https://github.com/imdario/mergo/issues/107

A separate named secret is used for openfaas/fn namespaces so
that a union merge can be performed using the name as a key.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>